Move command helpers under Console namespace.
This makes them available through the standalone Console package.
